requirements.txt
streamlit>=1.22.0
pandas>=1.3.0
numpy>=1.20.0
pdfplumber>=0.7.0
matplotlib>=3.5.0
seaborn>=1.0.0
rapidfuzz>=2.0.0
google-generativeai>=0.3.0
unidecode>=1.3.0
python-dateutil>=2.8.2
openpyxl>=3.0.9
